x/eightball/types/messages_scripture_signature_share_test.go
package types

import (
	"testing"

	"eightball/testutil/sample"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateScriptureSignatureShare_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateScriptureSignatureShare
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateScriptureSignatureShare{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateScriptureSignatureShare{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
